Move some functions from StrHelp to Str
- utf8_valid
- validate_utf8
- is_whitespace
- encode_utf8_char
