+++ runtime/core/Clownfish/CharBuf.c
@@ -30,7 +30,6 @@
 #include "Clownfish/Err.h"
 #include "Clownfish/String.h"
 #include "Clownfish/Util/Memory.h"
-#include "Clownfish/Util/StringHelper.h"
 #include "Clownfish/Class.h"
 
 // Append trusted UTF-8 to the CharBuf.
@@ -290,7 +289,7 @@ CB_Cat_Char_IMP(CharBuf *self, int32_t code_point) {
     size_t old_size = self->size;
     SI_add_grow_and_oversize(self, old_size, MAX_UTF8_BYTES);
     char *end = self->ptr + old_size;
-    size_t count = StrHelp_encode_utf8_char(code_point, (uint8_t*)end);
+    size_t count = Str_encode_utf8_char(code_point, (uint8_t*)end);
     self->size += count;
 }
