+++ runtime/core/Clownfish/Util/StringHelper.cfh
@@ -16,10 +16,6 @@
 
 parcel Clownfish;
 
-__C__
-#include "Clownfish/Err.h"
-__END_C__
-
 inert class Clownfish::Util::StringHelper nickname StrHelp {
 
     /* A table where the values indicate the number of bytes in a UTF-8
@@ -43,31 +39,6 @@ inert class Clownfish::Util::StringHelper nickname StrHelp {
     inert size_t
     to_base36(uint64_t value, void *buffer);
 
-    /** Return true if the string is valid UTF-8, false otherwise.
-     */
-    inert bool
-    utf8_valid(const char *ptr, size_t len);
-
-    /** Throws an error if the string isn't valid UTF-8.
-     */
-    inert void
-    validate_utf8(const char *text, size_t size, const char *file, int line,
-                  const char *func);
-
-    /** Returns true if the code point qualifies as Unicode whitespace.
-     */
-    inert bool
-    is_whitespace(int32_t code_point);
-
-    /** Encode a Unicode code point to a UTF-8 sequence.
-     *
-     * @param code_point A legal unicode code point.
-     * @param buffer Write buffer which must hold at least 4 bytes (the
-     * maximum legal length for a UTF-8 char).
-     */
-    inert uint32_t
-    encode_utf8_char(int32_t code_point, void *buffer);
-
     /** Return the first non-continuation byte before the supplied pointer.
      * If backtracking progresses beyond the supplied start, return NULL.
      */
@@ -76,17 +47,12 @@ inert class Clownfish::Util::StringHelper nickname StrHelp {
 }
 
 __C__
-#define CFISH_VALIDATE_UTF8(text, size) \
-    cfish_StrHelp_validate_utf8(text, size, \
-                                __FILE__, __LINE__, CFISH_ERR_FUNC_MACRO)
-
 /** The maximum number of bytes encoded by to_base36(), including the
  * terminating NULL.
  */
 #define cfish_StrHelp_MAX_BASE36_BYTES 14
 #ifdef CFISH_USE_SHORT_NAMES
   #define StrHelp_MAX_BASE36_BYTES cfish_StrHelp_MAX_BASE36_BYTES
-  #define VALIDATE_UTF8            CFISH_VALIDATE_UTF8
 #endif
 __END_C__
